+++ spec/integration/hanami/router/params_spec.rb
@@ -1,6 +1,9 @@
 # frozen_string_literal: true
 
 require "json"
+require "pathname"
+require "rack/builder"
+require "rack/multipart"
 require "hanami/middleware/body_parser"
 
 RSpec.describe "Params" do
@@ -71,6 +74,29 @@ RSpec.describe "Params" do
     end
   end
 
+  context "file upload" do
+    subject do
+      r = router
+
+      Rack::Builder.new do
+        use Hanami::Middleware::BodyParser, :form
+        run r
+      end
+    end
+
+    it "handles file upload" do
+      filename = "foo.xml"
+      env, contents = multipart_fixture(filename)
+
+      subject.call(env)
+
+      uploaded_file = env["router.params"].fetch(:file)
+
+      expect(uploaded_file.fetch(:filename)).to eq(filename)
+      expect(uploaded_file.fetch(:tempfile).read).to eq(contents)
+    end
+  end
+
   context "JSON payload" do
     subject do
       r = router
@@ -108,5 +134,26 @@ RSpec.describe "Params" do
 
       expect(env["router.params"]).to eq(id: expected)
     end
+  end
+
+  private
+
+  def multipart_fixture(filename, boundary = Rack::Multipart::MULTIPART_BOUNDARY)
+    path = fixture_path(filename)
+    file = Rack::Multipart::UploadedFile.new(path)
+    data = Rack::Multipart.build_multipart("file" => file)
+    env = Rack::MockRequest.env_for(
+      "/submit",
+      "CONTENT_TYPE" => "multipart/form-data; boundary=#{boundary}",
+      "CONTENT_LENGTH" => data.length.to_s,
+      method: "POST",
+      :input => StringIO.new(data)
+    )
+
+    [env, File.binread(path)]
+  end
+
+  def fixture_path(name)
+    Pathname.new(Dir.pwd).join("spec", "support", "fixtures", name).realpath
   end
 end