+++ spec/unit/hanami/router/version_spec.rb
@@ -2,6 +2,6 @@
 
 RSpec.describe "Hanami::Router::VERSION" do
   it "exposes version" do
-    expect(Hanami::Router::VERSION).to eq("2.0.0")
+    expect(Hanami::Router::VERSION).to eq("2.0.1")
   end
 end